# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Firestick API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/code/file": {
      "get": {
        "summary": "Get full file content",
        "parameters": [ { "name": "path", "in": "query", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "File content", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileContent" } } } }, "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }, "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    },
    "/api/code/chunk": {
      "get": {
        "summary": "Get file chunk by document id",
        "parameters": [ { "name": "docId", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1 }, "description": "Format: <path>#<type>:<start>-<end>" } ],
        "responses": { "200": { "description": "Chunk content", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileChunk" } } } }, "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }, "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "path": { "type": "string" },
          "requestId": { "type": "string", "description": "Correlation id; also returned in X-Request-Id header" },
          "status": { "type": "integer" },
          "error": { "type": "string" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
...
      "FileContent": {
        "type": "object",
        "properties": {
          "path": { "type": "string" },
          "content": { "type": "string" }
        }
      },
      "FileChunk": {
        "type": "object",
        "properties": {
          "path": { "type": "string" },
          "startLine": { "type": "integer" },
          "endLine": { "type": "integer" },
          "content": { "type": "string" }
        }
      },
...
    }
  }
}